Update README.md
This project focuses on the concept of Digital Twins in the context of the Internet of Things (IoT). A Digital Twin is a virtual representation of a physical object or system, designed to simulate its behavior in real-time. This project leverages IoT data to create and analyze Digital Twins, enabling enhanced monitoring, predictive maintenance, and performance optimization.